refactor: update import statements for AvailableAsset type in AvailableAssetsTable components
