+++ frontend/components/AgentWallet/BalancesAndAssets/AvailableAssetsTable.tsx
@@ -2,10 +2,10 @@ import { Flex, Image as AntdImage, TableColumnsType, Typography } from 'antd';
 
 import { Table } from '@/components/ui/Table';
 import { TokenSymbolConfigMap } from '@/constants/token';
+import { AvailableAsset } from '@/types/Wallet';
 import { formatNumber } from '@/utils/numberFormatters';
 
 import { useAgentWallet } from '../AgentWalletProvider';
-import { AvailableAsset } from '../types';
 
 const { Text } = Typography;
